feat: extract as reusable SimpleChart
