+++ src/pages/ThorChainLP/Pool/components/PoolChart.tsx
@@ -1,21 +1,18 @@
 import { Button, ButtonGroup, Center, Flex, Stack } from '@chakra-ui/react'
 import { thorchainAssetId } from '@shapeshiftoss/caip'
 import { useQuery } from '@tanstack/react-query'
-import type { ISeriesApi, SingleValueData, Time } from 'lightweight-charts'
-import { ColorType, createChart } from 'lightweight-charts'
-import { useCallback, useEffect, useMemo, useRef, useState } from 'react'
+import type { SingleValueData } from 'lightweight-charts'
+import { useCallback, useMemo, useState } from 'react'
 import { reactQueries } from 'react-queries'
 import type { Interval } from 'react-queries/queries/midgard'
+import { SimpleChart } from 'components/SimpleChart'
 import { fromThorBaseUnit } from 'lib/utils/thorchain'
 import type {
   MidgardSwapHistoryResponse,
   MidgardTvlHistoryResponse,
 } from 'lib/utils/thorchain/lp/types'
-import {
-  selectMarketDataByAssetIdUserCurrency,
-  selectSelectedCurrency,
-} from 'state/slices/selectors'
-import { store, useAppSelector } from 'state/store'
+import { selectMarketDataByAssetIdUserCurrency } from 'state/slices/selectors'
+import { useAppSelector } from 'state/store'
 
 const swapHistoryToChartData = (
   swapHistory: MidgardSwapHistoryResponse | undefined,
@@ -51,81 +48,12 @@ const tvlToChartData = (
     }
   })
 
-const currentLocale = window.navigator.languages[0]
-const selectedCurrency = selectSelectedCurrency(store.getState())
-const priceFormatter = Intl.NumberFormat(currentLocale, {
-  style: 'currency',
-  currency: selectedCurrency,
-  notation: 'compact',
-  compactDisplay: 'short',
-}).format
-
 // @ts-ignore we can't make this a partial record as we need to use this as a tuple to spread as useQuery params
 const INTERVAL_PARAMS_BY_INTERVAL: Record<Interval | 'all', [Interval, number]> = {
   day: ['hour', 24],
   week: ['hour', 24 * 7],
   month: ['day', 7 * 4],
   all: ['month', 24],
-}
-
-const backgroundColor = 'rgba(188, 214, 240, 0.04)'
-const lineColor = '#2962FF'
-const textColor = 'white'
-const areaTopColor = 'rgba(41, 98, 255, 0.5)'
-const areaBottomColor = 'rgba(41, 98, 255, 0.28)'
-
-type SimpleChartProps<T extends number | Time> = { data: SingleValueData<T>[] }
-
-export const SimpleChart = <T extends number | Time>({ data }: SimpleChartProps<T>) => {
-  const chartContainerRef = useRef<HTMLDivElement | null>(null)
-
-  useEffect(() => {
-    if (chartContainerRef.current && data) {
-      const chart = createChart(chartContainerRef.current, {
-        grid: {
-          vertLines: {
-            color: 'rgba(42, 46, 57, 0.5)',
-          },
-          horzLines: {
-            color: 'rgba(42, 46, 57, 0.5)',
-          },
-        },
-        layout: {
-          background: { type: ColorType.Solid, color: backgroundColor },
-          textColor,
-        },
-        width: chartContainerRef.current.offsetWidth,
-        height: chartContainerRef.current.offsetHeight,
-        localization: { priceFormatter },
-        timeScale: {
-          timeVisible: true,
-        },
-      })
-      const newSeries = chart.addAreaSeries({
-        lineColor,
-        topColor: areaTopColor,
-        bottomColor: areaBottomColor,
-      }) as unknown as ISeriesApi<'Area', T>
-      newSeries.setData(data)
-      chart.timeScale().fitContent()
-
-      const handleResize = () => {
-        chart.applyOptions({
-          width: chartContainerRef.current?.offsetWidth,
-          height: chartContainerRef.current?.offsetHeight,
-        })
-      }
-
-      window.addEventListener('resize', handleResize)
-
-      return () => {
-        window.removeEventListener('resize', handleResize)
-        chart.remove()
-      }
-    }
-  }, [data])
-
-  return <div ref={chartContainerRef} style={{ width: '100%', height: '500px' }} />
 }
 
 type PoolChartProps = {
